CSCETSIN-259: [ADD] moved map functions to different file (for testing)
